+++ src/testing/clean.cpp
@@ -35,6 +35,13 @@ void wps::clean( const eosio::name table, const std::optional<eosio::name> scope
             periods_itr = _periods.erase( periods_itr );
         }
         return;
+    // claims
+    } else if (table == "claims"_n) {
+        auto claims_itr = _claims.begin();
+        while ( claims_itr != _claims.end() ) {
+            claims_itr = _claims.erase( claims_itr );
+        }
+        return;
     // drafts
     } else if (table == "drafts"_n) {
         drafts_table _drafts( get_self(), scope->value );
@@ -43,19 +50,12 @@ void wps::clean( const eosio::name table, const std::optional<eosio::name> scope
             drafts_itr = _drafts.erase( drafts_itr );
         }
         return;
-    // transfers
-    } else if (table == "transfers"_n) {
-        auto transfers_itr = _transfers.begin();
-        while ( transfers_itr != _transfers.end() ) {
-            transfers_itr = _transfers.erase( transfers_itr );
-        }
-        return;
-    }
-    else if (table == "settings"_n) {
+    // settings
+    } else if (table == "settings"_n) {
         _settings.remove();
         return;
-    }
-    else if (table == "state"_n) {
+    // state
+    } else if (table == "state"_n) {
         _state.remove();
         return;
     }